feat: cleanup typescript runtime code and improve op order
Depend on https://github.com/dagger/dagger/pull/10094

This PR basically clean up the code by splitting responsibility.

TODO: cleanup generateClient code
