+++ sdk/typescript/runtime/main.go
@@ -4,47 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"path"
 	"path/filepath"
-	"slices"
-	"strings"
-
 	"typescript-sdk/internal/dagger"
 	"typescript-sdk/tsdistconsts"
-
-	"github.com/iancoleman/strcase"
-	"golang.org/x/mod/semver"
-)
-
-type SupportedTSRuntime string
-
-const (
-	Bun  SupportedTSRuntime = "bun"
-	Node SupportedTSRuntime = "node"
-	Deno SupportedTSRuntime = "deno"
-)
-
-type SupportedPackageManager string
-
-const (
-	Yarn        SupportedPackageManager = "yarn"
-	Pnpm        SupportedPackageManager = "pnpm"
-	Npm         SupportedPackageManager = "npm"
-	BunManager  SupportedPackageManager = "bun"
-	DenoManager SupportedPackageManager = "deno"
-)
-
-const (
-	PnpmDefaultVersion = "8.15.4"
-	YarnDefaultVersion = "1.22.22"
-	NpmDefaultVersion  = "10.7.0"
-)
-
-type SDKLibOrigin string
-
-const (
-	Bundle SDKLibOrigin = "bundle"
-	Local  SDKLibOrigin = "local"
 )
 
 func New(
@@ -55,43 +17,6 @@ func New(
 		SDKSourceDir: sdkSourceDir,
 		moduleConfig: &moduleConfig{},
 	}
-}
-
-type packageJSONConfig struct {
-	PackageManager string            `json:"packageManager"`
-	Dependencies   map[string]string `json:"dependencies"`
-	Dagger         *struct {
-		BaseImage string `json:"baseImage"`
-		Runtime   string `json:"runtime"`
-	} `json:"dagger"`
-}
-
-type denoJSONConfig struct {
-	Workspaces []string `json:"workspaces"`
-	Dagger     *struct {
-		BaseImage string `json:"baseImage"`
-	} `json:"dagger"`
-}
-
-type moduleConfig struct {
-	runtime        SupportedTSRuntime
-	runtimeVersion string
-
-	// Custom base image
-	image             string
-	packageJSONConfig *packageJSONConfig
-	denoJSONConfig    *denoJSONConfig
-
-	packageManager        SupportedPackageManager
-	packageManagerVersion string
-
-	source  *dagger.Directory
-	entries map[string]bool
-
-	name    string
-	subPath string
-
-	sdkLibOrigin SDKLibOrigin
 }
 
 type TypescriptSdk struct {
@@ -113,22 +38,35 @@ const (
 	codegenBinPath         = "/codegen"
 )
 
-// ModuleRuntime returns a container with the node or bun entrypoint ready to be called.
-func (t *TypescriptSdk) ModuleRuntime(ctx context.Context, modSource *dagger.ModuleSource, introspectionJSON *dagger.File) (*dagger.Container, error) {
+// ModuleRuntime implements the `ModuleRuntime` method from the SDK module interface.
+//
+// It returns a ready to call container with the correct node, bun or deno runtime setup.
+// On call, this will trigger the entrypoint that will either intropect and register the
+// module in the Dagger engine or execute a function of that module.
+//
+// The returned container has the codegen freshly generated and any necesary dependency
+// installed.
+func (t *TypescriptSdk) ModuleRuntime(
+	ctx context.Context,
+	modSource *dagger.ModuleSource,
+	introspectionJSON *dagger.File,
+) (*dagger.Container, error) {
 	err := t.analyzeModuleConfig(ctx, modSource)
 	if err != nil {
 		return nil, fmt.Errorf("failed to analyze module config: %w", err)
 	}
 
-	ctr, err := t.CodegenBase(ctx, modSource, introspectionJSON, true)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create codegen base: %w", err)
-	}
+	ctr := t.runtimeBaseContainer().
+		With(t.withConfiguredRuntimeEnvironment).
+		With(t.withGeneratedSDK(introspectionJSON)).
+		With(t.withSetupPackageManager).
+		With(t.withInstalledDependencies).
+		With(t.withUserSourceCode)
 
 	// Mount the entrypoint file
 	ctr = ctr.WithMountedFile(
 		t.moduleConfig.entrypointPath(),
-		ctr.Directory("/opt/module/bin").File(EntrypointExecutableFile),
+		entrypointFile(),
 	)
 
 	switch t.moduleConfig.runtime {
@@ -152,17 +90,34 @@ func (t *TypescriptSdk) ModuleRuntime(ctx context.Context, modSource *dagger.Mod
 	}
 }
 
-// Codegen returns the generated API client based on user's module
-func (t *TypescriptSdk) Codegen(ctx context.Context, modSource *dagger.ModuleSource, introspectionJSON *dagger.File) (*dagger.GeneratedCode, error) {
+// Codegen implements the `Codegen` method from the SDK module interface.
+//
+// It returns the generated API client based on user's module as well as
+// ignore directive regarding the generated content.
+func (t *TypescriptSdk) Codegen(
+	ctx context.Context,
+	modSource *dagger.ModuleSource,
+	introspectionJSON *dagger.File,
+) (*dagger.GeneratedCode, error) {
 	err := t.analyzeModuleConfig(ctx, modSource)
 	if err != nil {
 		return nil, fmt.Errorf("failed to analyze module config: %w", err)
 	}
 
-	// Get base container without dependencies installed.
-	ctr, err := t.CodegenBase(ctx, modSource, introspectionJSON, false)
+	ctr := t.runtimeBaseContainer().
+		With(t.withConfiguredRuntimeEnvironment).
+		With(t.withGeneratedSDK(introspectionJSON)).
+		With(t.withSetupPackageManager).
+		With(t.withGeneratedLockFile).
+		With(t.withUserSourceCode)
+
+	sourcesFiles, err := ctr.Directory(".").Glob(ctx, "src/**/*.ts")
 	if err != nil {
-		return nil, fmt.Errorf("failed to create codegen base: %w", err)
+		return nil, fmt.Errorf("failed to list user source files: %w", err)
+	}
+
+	if len(sourcesFiles) == 0 {
+		ctr = ctr.With(t.withInitTemplate)
 	}
 
 	// Extract codegen directory
@@ -313,651 +268,4 @@ func (t *TypescriptSdk) GenerateClient(
 	}
 
 	return dag.Directory().WithDirectory("/", ctr.Directory(workdirPath)), nil
-}
-
-// CodegenBase returns a Container containing the SDK from the engine container
-// and the user's code with a generated API based on what he did.
-func (t *TypescriptSdk) CodegenBase(ctx context.Context, modSource *dagger.ModuleSource, introspectionJSON *dagger.File, installDep bool) (*dagger.Container, error) {
-	base, err := t.Base()
-	if err != nil {
-		return nil, fmt.Errorf("failed to create codegen base: %w", err)
-	}
-
-	// Get a directory with the SDK sources installed and the generated client.
-	sdk := t.
-		addSDK().
-		WithFile("client.gen.ts", t.generateClient(base, introspectionJSON))
-
-	base = base.
-		// Add template directory
-		WithMountedDirectory("/opt/module", dag.CurrentModule().Source().Directory(".")).
-		// Mount user's module configuration (without sources) and the generated client in it.
-		WithDirectory(ModSourceDirPath,
-			dag.Directory().WithDirectory("/", modSource.ContextDirectory(), dagger.DirectoryWithDirectoryOpts{
-				Include: t.moduleConfigFiles(t.moduleConfig.subPath),
-			})).
-		WithDirectory(filepath.Join(t.moduleConfig.modulePath(), GenDir), sdk).
-		WithWorkdir(t.moduleConfig.modulePath())
-
-	base = t.configureModule(base)
-
-	// Generate the appropriate lock file
-	base, err = t.generateLockFile(base)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate lock file: %w", err)
-	}
-
-	// Install dependencies if needed.
-	if installDep {
-		base, err = t.installDependencies(base)
-		if err != nil {
-			return nil, fmt.Errorf("failed to install dependencies: %w", err)
-		}
-	}
-
-	// Add user's source files
-	base = base.WithDirectory(ModSourceDirPath,
-		dag.Directory().WithDirectory("/", modSource.ContextDirectory(), dagger.DirectoryWithDirectoryOpts{
-			// Include the rest of the user's module except config files to not override previous steps & SDKs.
-			Exclude: append(t.moduleConfigFiles(t.moduleConfig.subPath), filepath.Join(t.moduleConfig.subPath, "sdk")),
-		}),
-	)
-
-	base, err = t.addTemplate(ctx, base)
-	if err != nil {
-		return nil, fmt.Errorf("failed to add template: %w", err)
-	}
-
-	return base, nil
-}
-
-// Returns a list of files to include for module configs.
-func (t *TypescriptSdk) moduleConfigFiles(path string) []string {
-	modConfigFiles := []string{
-		"package.json",
-		"tsconfig.json",
-
-		// Workspaces files
-		"pnpm-workspace.yaml",
-		".yarnrc.yml",
-
-		// Lockfiles to include
-		"package-lock.json",
-		"yarn.lock",
-		"pnpm-lock.yaml",
-		"bun.lockb",
-		"bun.lock",
-
-		// Deno
-		"deno.json",
-		"deno.lock",
-	}
-
-	for i, file := range modConfigFiles {
-		modConfigFiles[i] = filepath.Join(path, file)
-	}
-
-	return modConfigFiles
-}
-
-// Base returns a Node or Bun container with cache setup for node package managers or bun
-func (t *TypescriptSdk) Base() (*dagger.Container, error) {
-	ctr := dag.Container().From(t.moduleConfig.image)
-
-	runtime := t.moduleConfig.runtime
-	version := t.moduleConfig.runtimeVersion
-
-	switch runtime {
-	case Bun:
-		return ctr.
-			WithoutEntrypoint().
-			WithMountedCache("/root/.bun/install/cache", dag.CacheVolume(fmt.Sprintf("mod-bun-cache-%s", tsdistconsts.DefaultBunVersion)), dagger.ContainerWithMountedCacheOpts{
-				Sharing: dagger.CacheSharingModePrivate,
-			}), nil
-	case Deno:
-		return ctr.
-			WithoutEntrypoint().
-			WithMountedCache("/root/.deno/cache", dag.CacheVolume(fmt.Sprintf("mod-deno-cache-%s", tsdistconsts.DefaultDenoVersion))), nil
-	case Node:
-		return ctr.
-			WithoutEntrypoint().
-			// Install default CA certificates and configure node to use them instead of its compiled in CA bundle.
-			// This enables use of custom CA certificates if configured in the dagger engine.
-			WithExec([]string{"apk", "add", "ca-certificates"}).
-			WithEnvVariable("NODE_OPTIONS", "--use-openssl-ca").
-			// Add cache volumes for npm, yarn and pnpm
-			WithMountedCache("/root/.npm", dag.CacheVolume(fmt.Sprintf("npm-cache-%s-%s", runtime, version))).
-			WithMountedCache("/root/.cache/yarn", dag.CacheVolume(fmt.Sprintf("yarn-cache-%s-%s", runtime, version))).
-			WithMountedCache("/root/.pnpm-store", dag.CacheVolume(fmt.Sprintf("pnpm-cache-%s-%s", runtime, version))).
-			// install tsx from its bundled location in the engine image
-			WithMountedDirectory("/usr/local/lib/node_modules/tsx", t.SDKSourceDir.Directory("/tsx_module")).
-			WithExec([]string{"ln", "-s", "/usr/local/lib/node_modules/tsx/dist/cli.mjs", "/usr/local/bin/tsx"}), nil
-
-	default:
-		return nil, fmt.Errorf("unknown runtime: %s", runtime)
-	}
-}
-
-// addTemplate adds the template files to the user's module if there's no
-// source files in the src directory.
-func (t *TypescriptSdk) addTemplate(ctx context.Context, ctr *dagger.Container) (*dagger.Container, error) {
-	name := t.moduleConfig.name
-
-	moduleFiles, err := ctr.Directory(".").Entries(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not list module source entries: %w", err)
-	}
-
-	// Check if there's a src directory and creates an empty directory if it doesn't exist.
-	if !slices.Contains(moduleFiles, "src") {
-		ctr = ctr.WithDirectory("src", dag.Directory())
-	}
-
-	// Get the list of files in the src directory.
-	moduleSourceFiles, err := ctr.Directory("src").Entries(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not list module source entries: %w", err)
-	}
-
-	// Check if there's a src directory with .ts files in it.
-	// If not, add the template file and replace QuickStart with the module name
-	if !slices.ContainsFunc(moduleSourceFiles, func(s string) bool {
-		return path.Ext(s) == ".ts"
-	}) {
-		return ctr.
-			WithDirectory("src", ctr.Directory("/opt/module/template/src"), dagger.ContainerWithDirectoryOpts{Include: []string{"*.ts"}}).
-			WithExec([]string{"sed", "-i", "-e", fmt.Sprintf("s/QuickStart/%s/g", strcase.ToCamel(name)), "src/index.ts"}), nil
-	}
-
-	return ctr, nil
-}
-
-// setupModule configure the user's module.
-//
-// If the user's module has a package.json file, it will run the
-// __tsconfig.updator.ts script in order to add dagger to the tsconfig path so
-// the editor can give type hints and auto completion.
-// Otherwise, it will copy the template config files into the user's module directory.
-//
-// If there's no src directory or no typescript files in it, it will create one
-// and copy the template index.ts file in it.
-func (t *TypescriptSdk) configureModule(ctr *dagger.Container) *dagger.Container {
-	runtime := t.moduleConfig.runtime
-
-	// If there's a package.json in bun & node, run the tsconfig updator script and install the genDir.
-	templateSetup := func() *dagger.Container {
-		return ctr.WithDirectory(".",
-			ctr.Directory("/opt/module/template"), dagger.ContainerWithDirectoryOpts{Include: []string{"*.json"}})
-	}
-
-	switch runtime {
-	case Bun:
-		if t.moduleConfig.packageJSONConfig == nil {
-			return templateSetup()
-		}
-
-		return ctr.
-			WithExec([]string{"bun", "/opt/module/bin/__tsconfig.updator.ts", fmt.Sprintf("--sdk-lib-origin=%s", t.moduleConfig.sdkLibOrigin)}).
-			WithFile("package.json", t.configurePackageJSON(ctr.File("package.json")))
-	case Node:
-		if t.moduleConfig.packageJSONConfig == nil {
-			return templateSetup()
-		}
-
-		return ctr.
-			WithExec([]string{"tsx", "/opt/module/bin/__tsconfig.updator.ts", fmt.Sprintf("--sdk-lib-origin=%s", t.moduleConfig.sdkLibOrigin)}).
-			WithFile("package.json", t.configurePackageJSON(ctr.File("package.json")))
-	case Deno:
-		return ctr.WithExec([]string{"deno", "run", "-A", "/opt/module/bin/__deno_config_updator.ts", fmt.Sprintf("--sdk-lib-origin=%s", t.moduleConfig.sdkLibOrigin)})
-	default:
-		return ctr
-	}
-}
-
-// Update the user's package.json with required dependencies to use the Typescript SDK.
-func (t *TypescriptSdk) configurePackageJSON(file *dagger.File) *dagger.File {
-	ctr := dag.
-		Container().
-		From(tsdistconsts.DefaultNodeImageRef).
-		WithDirectory("/src", dag.Directory().WithFile("package.json", file)).
-		WithWorkdir("/src").
-		WithExec([]string{"npm", "pkg", "set", "type=module"})
-
-	if t.moduleConfig.packageJSONConfig != nil {
-		_, ok := t.moduleConfig.packageJSONConfig.Dependencies["typescript"]
-		if !ok {
-			ctr = ctr.WithExec([]string{"npm", "pkg", "set", "dependencies.typescript=^5.5.4"})
-		}
-	}
-
-	return ctr.File("/src/package.json")
-}
-
-// addSDK returns a directory with the SDK sources.
-func (t *TypescriptSdk) addSDK() *dagger.Directory {
-	switch t.moduleConfig.sdkLibOrigin {
-	case Bundle:
-		return t.SDKSourceDir.
-			Directory("/bundled_lib").
-			WithDirectory("/", dag.CurrentModule().Source().Directory("bundled_static_export/module"))
-	case Local:
-		return t.SDKSourceDir.
-			WithoutDirectory("codegen").
-			WithoutDirectory("runtime").
-			WithoutDirectory("tsx_module").
-			WithoutDirectory("bundled_Lib").
-			WithoutDirectory("src/provisioning")
-	default:
-		// This should never happen.
-		return dag.Directory()
-	}
-}
-
-// generateClient uses the given container to generate the client code.
-func (t *TypescriptSdk) generateClient(ctr *dagger.Container, introspectionJSON *dagger.File) *dagger.File {
-	codegenArgs := []string{
-		codegenBinPath,
-		"--lang", "typescript",
-		"--output", ModSourceDirPath,
-		"--module-name", t.moduleConfig.name,
-		"--module-source-path", t.moduleConfig.modulePath(),
-		"--introspection-json-path", schemaPath,
-	}
-
-	if t.moduleConfig.sdkLibOrigin == Bundle {
-		codegenArgs = append(codegenArgs, "--bundle")
-	}
-
-	return ctr.
-		// Add dagger codegen binary.
-		WithMountedFile(codegenBinPath, t.SDKSourceDir.File("/codegen")).
-		// Mount the introspection file.
-		WithMountedFile(schemaPath, introspectionJSON).
-		// Execute the code generator using the given introspection file.
-		WithExec(codegenArgs, dagger.ContainerWithExecOpts{
-			ExperimentalPrivilegedNesting: true,
-		}).
-		// Return the generated code directory.
-		Directory(t.moduleConfig.sdkPath()).
-		File("/src/api/client.gen.ts")
-}
-
-// detectSDKLibOrigin return the SDK library config based on the user's module config.
-// For Node & Bun:
-// - if the package.json has `dependencies[@dagger.io/dagger]=./sdk` -> Return Local
-// - else -> Return Bundle
-// For Deno:
-// - if deno.json has `workspaces` with `./sdk` in it -> Return Local
-// - else -> Return Bundle
-// Return Bundle by default since it's more performant.
-func (t *TypescriptSdk) detectSDKLibOrigin() (SDKLibOrigin, error) {
-	runtime := t.moduleConfig.runtime
-
-	switch runtime {
-	case Node, Bun:
-		if t.moduleConfig.packageJSONConfig != nil && t.moduleConfig.packageJSONConfig.Dependencies["@dagger.io/dagger"] == "./sdk" {
-			return Local, nil
-		}
-
-		return Bundle, nil
-	case Deno:
-		if t.moduleConfig.denoJSONConfig != nil && slices.Contains(t.moduleConfig.denoJSONConfig.Workspaces, "./sdk") {
-			return Local, nil
-		}
-
-		return Bundle, nil
-	default:
-		return Bundle, fmt.Errorf("unknown runtime: %s", runtime)
-	}
-}
-
-// detectBaseImageRef return the base image ref of the runtime
-// based on the user's module config.
-//
-// If set in the `dagger.baseImage` field of the module's package.json, the
-// runtime use that ref.
-// If not set, it return the default base image ref based on the detected runtime and
-// it's version.
-//
-// Note: This function must be called after `detectRuntime`.
-func (t *TypescriptSdk) detectBaseImageRef() (string, error) {
-	runtime := t.moduleConfig.runtime
-	version := t.moduleConfig.runtimeVersion
-
-	if t.moduleConfig.packageJSONConfig != nil && t.moduleConfig.packageJSONConfig.Dagger != nil {
-		value := t.moduleConfig.packageJSONConfig.Dagger.BaseImage
-		if value != "" {
-			return value, nil
-		}
-	}
-
-	switch runtime {
-	case Bun:
-		if version != "" {
-			return fmt.Sprintf("oven/%s:%s-alpine", Bun, version), nil
-		}
-
-		return tsdistconsts.DefaultBunImageRef, nil
-	case Node:
-		if version != "" {
-			return fmt.Sprintf("%s:%s-alpine", Node, version), nil
-		}
-
-		return tsdistconsts.DefaultNodeImageRef, nil
-	case Deno:
-		if t.moduleConfig.denoJSONConfig != nil && t.moduleConfig.denoJSONConfig.Dagger != nil {
-			value := t.moduleConfig.denoJSONConfig.Dagger.BaseImage
-			if value != "" {
-				return value, nil
-			}
-		}
-
-		return tsdistconsts.DefaultDenoImageRef, nil
-	default:
-		return "", fmt.Errorf("unknown runtime: %q", runtime)
-	}
-}
-
-// DetectRuntime returns the runtime(bun or node) detected for the user's module
-// If a runtime is specfied inside the package.json, it will be used.
-// If a package-lock.json, yarn.lock, or pnpm-lock.yaml is present, node will be used.
-// If a bun.lock or bun.lockb is present, bun will be used.
-// If none of the above is present, node will be used.
-//
-// If the runtime is detected and pinned to a specific version, it will also return the pinned version.
-func (t *TypescriptSdk) detectRuntime() error {
-	// If we find a package.json, we check if the runtime is specified in `dagger.runtime` field.
-	if t.moduleConfig.packageJSONConfig != nil && t.moduleConfig.packageJSONConfig.Dagger != nil {
-		value := t.moduleConfig.packageJSONConfig.Dagger.Runtime
-		if value != "" {
-			// Retrieve the runtime and version from the value (e.g., node@lts, bun@1)
-			// If version isn't specified, version will be an empty string and only the runtime will be used in Base.
-			runtime, version, _ := strings.Cut(value, "@")
-
-			switch runtime := SupportedTSRuntime(runtime); runtime {
-			case Bun, Node:
-				t.moduleConfig.runtime = runtime
-				t.moduleConfig.runtimeVersion = version
-
-				return nil
-			default:
-				return fmt.Errorf("unsupported runtime %q", runtime)
-			}
-		}
-	}
-
-	// Try to detect runtime from lock files
-	if t.moduleConfig.hasFile("bun.lockb") || t.moduleConfig.hasFile("bun.lock") {
-		t.moduleConfig.runtime = Bun
-
-		return nil
-	}
-
-	if t.moduleConfig.hasFile("package-lock.json") ||
-		t.moduleConfig.hasFile("yarn.lock") ||
-		t.moduleConfig.hasFile("pnpm-lock.yaml") {
-		t.moduleConfig.runtime = Node
-
-		return nil
-	}
-
-	if t.moduleConfig.hasFile("deno.json") ||
-		t.moduleConfig.hasFile("deno.lock") {
-		t.moduleConfig.runtime = Deno
-
-		return nil
-	}
-
-	return nil
-}
-
-// detectPackageManager detects the package manager from the user's module.
-// If the package.json file has a field "packageManager", it will use that to
-// determine the package manager to use. Otherwise, it will use the default
-// package manager based on the lock files present in the module.
-//
-// If none of the above works, it will use yarn.
-//
-// Except if the package.json has an invalid value in field "packageManager", this
-// function should never return an error.
-func (t *TypescriptSdk) detectPackageManager() (SupportedPackageManager, string, error) {
-	// If the runtime is Bun, we should use BunManager
-	if t.moduleConfig.runtime == Bun {
-		return BunManager, "", nil
-	}
-
-	// If the runtime is deno, we should use the DenoManager
-	if t.moduleConfig.runtime == Deno {
-		return DenoManager, "", nil
-	}
-
-	// If we find a package.json, we check if the packageManager is specified in `packageManager` field.
-	if t.moduleConfig.packageJSONConfig != nil {
-		value := t.moduleConfig.packageJSONConfig.PackageManager
-		if value != "" {
-			// Retrieve the package manager and version from the value (e.g., yarn@4.2.0, pnpm@8.5.1)
-			packageManager, version, _ := strings.Cut(value, "@")
-
-			if version == "" {
-				return "", "", fmt.Errorf("packageManager version is missing, please add it to your package.json")
-			}
-
-			switch SupportedPackageManager(packageManager) {
-			case Yarn, Pnpm, Npm:
-				return SupportedPackageManager(packageManager), version, nil
-			default:
-				return "", "", fmt.Errorf("detected unknown package manager: %s", packageManager)
-			}
-		}
-	}
-
-	// Bun can additionally output a yarn.lock file, so we need to check for Bun before Yarn.
-	if t.moduleConfig.hasFile("bun.lockb") || t.moduleConfig.hasFile("bun.lock") {
-		return BunManager, "", nil
-	}
-
-	if t.moduleConfig.hasFile("package-lock.json") {
-		return Npm, NpmDefaultVersion, nil
-	}
-
-	if t.moduleConfig.hasFile("yarn.lock") {
-		return Yarn, YarnDefaultVersion, nil
-	}
-
-	if t.moduleConfig.hasFile("pnpm-lock.yaml") {
-		return Pnpm, PnpmDefaultVersion, nil
-	}
-
-	// Default to yarn
-	return Yarn, YarnDefaultVersion, nil
-}
-
-// generateLockFile generate a lock file for the matching package manager.
-func (t *TypescriptSdk) generateLockFile(ctr *dagger.Container) (*dagger.Container, error) {
-	packageManager := t.moduleConfig.packageManager
-	version := t.moduleConfig.packageManagerVersion
-
-	switch packageManager {
-	case Yarn:
-		// Enable corepack
-		// NOTE: this incidentally fetches and installs node modules, maybe we could install yarn some other way?
-		ctr = ctr.
-			WithExec([]string{"corepack", "enable"}).
-			WithExec([]string{"corepack", "use", fmt.Sprintf("yarn@%s", version)})
-
-		// Install dependencies and extract the lockfile
-		file := ctr.
-			WithExec([]string{"yarn", "install", "--mode", "update-lockfile"}).File("yarn.lock")
-
-		// We use node-modules linker for yarn >= v3 because it's not working with pnp.
-		if semver.Compare(fmt.Sprintf("v%s", t.moduleConfig.packageManagerVersion), "v3.0.0") >= 0 {
-			ctr = ctr.WithNewFile(".yarnrc.yml", `nodeLinker: node-modules`)
-		}
-
-		// Sadly, yarn < v3 doesn't support generating a lockfile without installing the dependencies.
-		// So we use npm to generate the lockfile and then import it into yarn.
-		return ctr.WithFile("yarn.lock", file), nil
-	case Pnpm:
-		ctr = ctr.WithExec([]string{"npm", "install", "-g", fmt.Sprintf("pnpm@%s", version)})
-
-		if !t.moduleConfig.hasFile("pnpm-workspace.yaml") && t.moduleConfig.sdkLibOrigin == Local {
-			ctr = ctr.
-				WithNewFile("pnpm-workspace.yaml", `packages:
-  - './sdk'
-			`)
-		}
-
-		return ctr.WithExec([]string{"pnpm", "install", "--lockfile-only"}), nil
-	case Npm:
-		return ctr.
-			WithExec([]string{"npm", "install", "-g", fmt.Sprintf("npm@%s", version)}).
-			WithExec([]string{"npm", "install", "--package-lock-only"}), nil
-	case BunManager:
-		return ctr.
-			WithExec([]string{"bun", "install", "--no-verify", "--no-progress"}), nil
-	case DenoManager:
-		return ctr, nil
-	default:
-		return nil, fmt.Errorf("detected unknown package manager: %s", packageManager)
-	}
-}
-
-// installDependencies installs the dependencies using the detected package manager.
-func (t *TypescriptSdk) installDependencies(ctr *dagger.Container) (*dagger.Container, error) {
-	switch t.moduleConfig.packageManager {
-	case Yarn:
-		if semver.Compare(fmt.Sprintf("v%s", t.moduleConfig.packageManagerVersion), "v3.0.0") <= 0 {
-			return ctr.
-				WithExec([]string{"yarn", "install", "--frozen-lockfile"}), nil
-		}
-
-		return ctr.WithExec([]string{"yarn", "install", "--immutable"}), nil
-	case Pnpm:
-		return ctr.
-			WithExec([]string{"pnpm", "install", "--frozen-lockfile", "--shamefully-hoist=true"}), nil
-	case Npm:
-		return ctr.
-			WithExec([]string{"npm", "ci"}), nil
-	case BunManager:
-		return ctr.
-			WithExec([]string{"bun", "install", "--no-verify", "--no-progress"}), nil
-	case DenoManager:
-		return ctr.
-			WithExec([]string{"deno", "install"}), nil
-	default:
-		return nil, fmt.Errorf("detected unknown package manager: %s", t.moduleConfig.packageManager)
-	}
-}
-
-// analyzeModuleConfig analyzes the module config and populates the moduleConfig field.
-//
-// It detect the module name, source subpath, runtime, package manager and their versions.
-// It also populates the moduleConfig.entries map with the list of files present in the module source.
-//
-// It's a utility function that should be called before calling any other exposed function in this module.
-func (t *TypescriptSdk) analyzeModuleConfig(ctx context.Context, modSource *dagger.ModuleSource) (err error) {
-	if t.moduleConfig == nil {
-		t.moduleConfig = &moduleConfig{
-			entries:           make(map[string]bool),
-			runtime:           Node,
-			packageJSONConfig: nil,
-			sdkLibOrigin:      Bundle,
-		}
-	}
-
-	t.moduleConfig.name, err = modSource.ModuleOriginalName(ctx)
-	if err != nil {
-		return fmt.Errorf("could not load module config name: %w", err)
-	}
-
-	t.moduleConfig.subPath, err = modSource.SourceSubpath(ctx)
-	if err != nil {
-		return fmt.Errorf("could not load module config source subpath: %w", err)
-	}
-
-	// If a first init, there will be no directory, so we ignore the error here.
-	// We also only include package.json & lockfiles to benefit from caching.
-	t.moduleConfig.source = modSource.ContextDirectory().Directory(t.moduleConfig.subPath)
-	configEntries, err := dag.Directory().WithDirectory(".", t.moduleConfig.source, dagger.DirectoryWithDirectoryOpts{
-		Include: t.moduleConfigFiles("."),
-	}).Entries(ctx)
-	if err == nil {
-		for _, entry := range configEntries {
-			t.moduleConfig.entries[entry] = true
-		}
-	}
-
-	if t.moduleConfig.hasFile("package.json") {
-		var packageJSONConfig packageJSONConfig
-
-		content, err := t.moduleConfig.source.File("package.json").Contents(ctx)
-		if err != nil {
-			return fmt.Errorf("failed to read package.json: %w", err)
-		}
-
-		if err := json.Unmarshal([]byte(content), &packageJSONConfig); err != nil {
-			return fmt.Errorf("failed to unmarshal package.json: %w", err)
-		}
-
-		t.moduleConfig.packageJSONConfig = &packageJSONConfig
-	}
-
-	if t.moduleConfig.hasFile("deno.json") {
-		var denoJSONConfig denoJSONConfig
-
-		content, err := t.moduleConfig.source.File("deno.json").Contents(ctx)
-		if err != nil {
-			return fmt.Errorf("failed to read deno.json: %w", err)
-		}
-
-		if err := json.Unmarshal([]byte(content), &denoJSONConfig); err != nil {
-			return fmt.Errorf("failed to unmarshal deno.json: %w", err)
-		}
-
-		t.moduleConfig.denoJSONConfig = &denoJSONConfig
-	}
-
-	if err := t.detectRuntime(); err != nil {
-		return fmt.Errorf("failed to detect module runtime: %w", err)
-	}
-
-	t.moduleConfig.packageManager, t.moduleConfig.packageManagerVersion, err = t.detectPackageManager()
-	if err != nil {
-		return fmt.Errorf("failed to detect package manager: %w", err)
-	}
-
-	t.moduleConfig.image, err = t.detectBaseImageRef()
-	if err != nil {
-		return fmt.Errorf("failed to detect base image ref: %w", err)
-	}
-
-	t.moduleConfig.sdkLibOrigin, err = t.detectSDKLibOrigin()
-	if err != nil {
-		return fmt.Errorf("failed to detect sdk lib origin: %w", err)
-	}
-
-	return nil
-}
-
-// Return true if the file is present in the module source.
-func (c *moduleConfig) hasFile(name string) bool {
-	_, ok := c.entries[name]
-
-	return ok
-}
-
-func (c *moduleConfig) modulePath() string {
-	return filepath.Join(ModSourceDirPath, c.subPath)
-}
-
-func (c *moduleConfig) sdkPath() string {
-	return filepath.Join(c.modulePath(), GenDir)
-}
-
-func (c *moduleConfig) entrypointPath() string {
-	return filepath.Join(ModSourceDirPath, c.subPath, SrcDir, EntrypointExecutableFile)
-}
-
-func (c *moduleConfig) tsConfigPath() string {
-	return filepath.Join(ModSourceDirPath, c.subPath, "tsconfig.json")
 }
